move modules into module/; remove robot-data/
