+++ shared-data/js/modules.js
@@ -1,5 +1,5 @@
 // @flow
-import moduleSpecs from '../robot-data/moduleSpecs.json'
+import moduleSpecs from '../module/schemas/1.json'
 
 export type ModuleType = 'magdeck' | 'tempdeck'
